Test build / warning fixes for iOS
googletest's EXPECT_DEATH is not supported on iOS, so disable those tests.
Also fix a couple of compiler warnings that were raised.

Issue: #126

+++ src/thread.cpp
@@ -226,7 +226,7 @@ void Thread::setName(const char* fmt, ...) {
 }
 
 unsigned int Thread::numLogicalCPUs() {
-  return sysconf(_SC_NPROCESSORS_ONLN);
+  return static_cast<unsigned int>(sysconf(_SC_NPROCESSORS_ONLN));
 }
 
 #endif  // OS
